Move all tuners to sub-modules
For each tuner, created a sub-module that contains at least:

- config.py for config stuff
- model.py for the actual model/encoder/embedding
- __init__.py so that imports are preserved

Then, when there was a need, further files were created, like layer.py
or utils.py.

For some existing modules, the license comment of the top of the file
was missing, I always added it.
