+++ src/peft/tuners/ia3/model.py
@@ -12,117 +12,31 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import re
 import warnings
-from dataclasses import asdict, dataclass, field
+from dataclasses import asdict
 from enum import Enum
-from typing import List, Optional, Union
 
+import bitsandbytes as bnb
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
 from transformers.pytorch_utils import Conv1D
 
-from ..config import PeftConfig
-from ..import_utils import is_bnb_available
-from ..utils import (
+from peft.import_utils import is_bnb_available
+from peft.tuners.tuners_utils import BaseTuner
+from peft.utils import (
     TRANSFORMERS_MODELS_TO_IA3_FEEDFORWARD_MODULES_MAPPING,
     TRANSFORMERS_MODELS_TO_IA3_TARGET_MODULES_MAPPING,
     ModulesToSaveWrapper,
-    PeftType,
     _get_submodules,
     _is_valid_match,
-    transpose,
 )
-from .tuners_utils import BaseTuner, BaseTunerLayer
+
+from .layer import IA3Layer, Linear
 
 
 if is_bnb_available():
-    import bitsandbytes as bnb
-
-
-@dataclass
-class IA3Config(PeftConfig):
-    """
-    This is the configuration class to store the configuration of a [`IA3Model`].
-
-    Args:
-        target_modules (`Union[List[str],str]`): The names of the modules to apply (IA)^3 to.
-        feedforward_modules (`Union[List[str],str]`): The names of the modules to be treated as feedforward modules
-        as in the original paper.
-        fan_in_fan_out (`bool`): Set this to True if the layer to replace stores weight like (fan_in, fan_out).
-        For example, gpt-2 uses `Conv1D` which stores weights like (fan_in, fan_out) and hence this should be set to `True`.:
-        modules_to_save (`List[str]`):List of modules apart from (IA)^3 layers to be set as trainable
-            and saved in the final checkpoint.
-        init_ia3_weights (`bool`): Whether to initialize the vectors in the (IA)^3 layers, defaults to `True`.
-    """
-
-    target_modules: Optional[Union[List[str], str]] = field(
-        default=None,
-        metadata={
-            "help": "List of module names or regex expression of the module names to replace with ia3."
-            "For example, ['q', 'v'] or '.*decoder.*(SelfAttention|EncDecAttention).*(q|v)$' "
-        },
-    )
-    feedforward_modules: Optional[Union[List[str], str]] = field(
-        default=None,
-        metadata={
-            "help": "List of module names or a regex expression of module names which are feedforward"
-            "For example, ['output.dense']"
-        },
-    )
-    fan_in_fan_out: bool = field(
-        default=False,
-        metadata={"help": "Set this to True if the layer to replace stores weight like (fan_in, fan_out)"},
-    )
-    modules_to_save: Optional[List[str]] = field(
-        default=None,
-        metadata={
-            "help": "List of modules apart from (IA)^3 layers to be set as trainable and saved in the final checkpoint. "
-            "For example, in Sequence Classification or Token Classification tasks, "
-            "the final layer `classifier/score` are randomly initialized and as such need to be trainable and saved."
-        },
-    )
-    init_ia3_weights: bool = field(
-        default=True,
-        metadata={"help": "Whether to initialize the vectors in the (IA)^3 layers."},
-    )
-
-    def __post_init__(self):
-        self.peft_type = PeftType.IA3
-
-
-class IA3Layer(BaseTunerLayer):
-    def __init__(
-        self,
-        in_features: int,
-        out_features: int,
-        is_feedforward: bool,
-    ):
-        self.scaling = {}
-        self.ia3_l = nn.ParameterDict({})
-        # Mark the weight as unmerged
-        self.merged = False
-        self.disable_adapters = False
-        self.in_features = in_features
-        self.out_features = out_features
-        self.is_feedforward = is_feedforward
-
-    def update_layer(self, adapter_name, init_ia3_weights):
-        # Actual trainable parameters
-        if self.is_feedforward:
-            weight = torch.randn((1, self.in_features))
-        else:
-            weight = torch.randn((self.out_features, 1))
-        self.ia3_l[adapter_name] = nn.Parameter(weight)
-        if init_ia3_weights:
-            self.reset_ia3_parameters(adapter_name)
-        self.to(self.weight.device)
-
-    def reset_ia3_parameters(self, adapter_name):
-        if adapter_name in self.ia3_l.keys():
-            # initialize learned vector with torch.ones
-            nn.init.constant_(self.ia3_l[adapter_name], 1.0)
+    from .bnb import Linear8bitLt
 
 
 class IA3Model(BaseTuner):
@@ -358,146 +272,3 @@ class IA3Model(BaseTuner):
                 setattr(parent, target_name, target.modules_to_save[target.active_adapter])
 
         return self.model
-
-
-class Linear(nn.Linear, IA3Layer):
-    # (IA)^3 implemented in a dense layer
-    def __init__(
-        self,
-        adapter_name: str,
-        in_features: int,
-        out_features: int,
-        fan_in_fan_out: bool = False,  # Set this to True if the layer to replace stores weight like (fan_in, fan_out)
-        is_feedforward: bool = False,  # Set to True if the layer is treated as a feedforward layer
-        **kwargs,
-    ) -> None:
-        init_ia3_weights = kwargs.pop("init_ia3_weights", True)
-
-        nn.Linear.__init__(self, in_features, out_features, **kwargs)
-        IA3Layer.__init__(self, in_features=in_features, out_features=out_features, is_feedforward=is_feedforward)
-        # Freezing the pre-trained weight matrix
-        self.weight.requires_grad = False
-
-        self.fan_in_fan_out = fan_in_fan_out
-        if fan_in_fan_out:
-            self.weight.data = self.weight.data.T
-
-        nn.Linear.reset_parameters(self)
-        self.update_layer(adapter_name, init_ia3_weights)
-        self.active_adapter = adapter_name
-
-        self.is_feedforward = is_feedforward
-
-    def merge(self) -> None:
-        if self.active_adapter not in self.ia3_l.keys():
-            return
-        if self.merged:
-            warnings.warn("Already merged. Nothing to do.")
-            return
-
-        self.weight = transpose(self.weight, self.fan_in_fan_out)
-        self.weight.data = torch.mul(self.weight.data, self.ia3_l[self.active_adapter].data)
-        self.weight = transpose(self.weight, self.fan_in_fan_out)
-
-        self.merged = True
-
-    def unmerge(self) -> None:
-        if self.active_adapter not in self.ia3_l.keys():
-            return
-        if not self.merged:
-            warnings.warn("Already unmerged. Nothing to do.")
-            return
-
-        warnings.warn("Unmerge result can be inaccurate for (IA)^3.")
-        self.weight = transpose(self.weight, self.fan_in_fan_out)
-        # divide by (IA)^3 vector. Add tolerace to avoid division by zero
-        self.weight.data = torch.div(self.weight.data, self.ia3_l[self.active_adapter].data + 1e-8)
-        self.weight = transpose(self.weight, self.fan_in_fan_out)
-
-        self.merged = False
-
-    def _linear(self, input: torch.Tensor) -> torch.Tensor:
-        return F.linear(input, transpose(self.weight, self.fan_in_fan_out), bias=self.bias)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        if self.active_adapter not in self.ia3_l.keys():
-            return self._linear(x)
-
-        previous_dtype = x.dtype
-
-        if self.disable_adapters:
-            if self.merged:
-                self.unmerge()
-            result = self._linear(x)
-        elif self.merged:
-            result = self._linear(x)
-        else:
-            dtype = self.ia3_l[self.active_adapter].dtype
-            ia3_scaling = self.ia3_l[self.active_adapter].flatten()
-            if self.is_feedforward:
-                x = x.to(dtype)
-                # TODO: self.weight.dtype can be != self.ia3_l[self.active_adapter].dtype
-                # e.g. bf16 vs fp32. Is that okay?
-                interm = (x * ia3_scaling).to(self.weight.dtype)
-                result = self._linear(interm)
-            else:
-                result = self._linear(x)
-                result = result.to(dtype) * ia3_scaling
-
-        result = result.to(previous_dtype)
-        return result
-
-
-if is_bnb_available():
-
-    class Linear8bitLt(bnb.nn.Linear8bitLt, IA3Layer):
-        # (IA)^3 implemented in a dense layer
-        def __init__(
-            self,
-            adapter_name,
-            in_features,
-            out_features,
-            is_feedforward,
-            **kwargs,
-        ) -> None:
-            bnb.nn.Linear8bitLt.__init__(
-                self,
-                in_features,
-                out_features,
-                bias=kwargs.get("bias", True),
-                has_fp16_weights=kwargs.get("has_fp16_weights", True),
-                memory_efficient_backward=kwargs.get("memory_efficient_backward", False),
-                threshold=kwargs.get("threshold", 0.0),
-                index=kwargs.get("index", None),
-            )
-            IA3Layer.__init__(self, in_features=in_features, out_features=out_features, is_feedforward=is_feedforward)
-
-            # Freezing the pre-trained weight matrix
-            self.weight.requires_grad = False
-
-            init_ia3_weights = kwargs.pop("init_ia3_weights", True)
-            self.update_layer(adapter_name, init_ia3_weights)
-            self.active_adapter = adapter_name
-            self.is_feedforward = is_feedforward
-
-        def forward(self, x: torch.Tensor) -> torch.Tensor:
-            if self.disable_adapters or (self.active_adapter not in self.ia3_l.keys()):
-                return super().forward(x)
-
-            requires_conversion = (not torch.is_autocast_enabled()) and (x.dtype != torch.float32)
-            if requires_conversion:
-                x = x.float()
-
-            ia3_scaling = self.ia3_l[self.active_adapter].flatten()
-            if self.is_feedforward:
-                result = super().forward(x * ia3_scaling)
-                expected_dtype = result.dtype
-            else:
-                result = super().forward(x)
-                expected_dtype = result.dtype
-                result = result * ia3_scaling
-
-            if requires_conversion:
-                result = result.to(expected_dtype)
-
-            return result
