+++ src/peft/tuners/lora/layer.py
@@ -15,19 +15,14 @@
 
 import math
 import warnings
-from typing import Tuple, Union
+from typing import Optional, Tuple, Union
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from peft.import_utils import is_bnb_4bit_available, is_bnb_available
 from peft.tuners.tuners_utils import BaseTunerLayer
 from peft.utils.other import transpose
-
-
-if is_bnb_available():
-    import bitsandbytes as bnb
 
 
 class LoraLayer(BaseTunerLayer):
@@ -398,188 +393,3 @@ class Conv2d(nn.Conv2d, LoraLayer):
 
         result = result.to(previous_dtype)
         return result
-
-
-if is_bnb_available():
-
-    class Linear8bitLt(bnb.nn.Linear8bitLt, LoraLayer):
-        # Lora implemented in a dense layer
-        def __init__(
-            self,
-            adapter_name,
-            in_features,
-            out_features,
-            r: int = 0,
-            lora_alpha: int = 1,
-            lora_dropout: float = 0.0,
-            **kwargs,
-        ) -> None:
-            bnb.nn.Linear8bitLt.__init__(
-                self,
-                in_features,
-                out_features,
-                bias=kwargs.get("bias", True),
-                has_fp16_weights=kwargs.get("has_fp16_weights", True),
-                memory_efficient_backward=kwargs.get("memory_efficient_backward", False),
-                threshold=kwargs.get("threshold", 0.0),
-                index=kwargs.get("index", None),
-            )
-            LoraLayer.__init__(self, in_features=in_features, out_features=out_features)
-
-            # Freezing the pre-trained weight matrix
-            self.weight.requires_grad = False
-            init_lora_weights = kwargs.pop("init_lora_weights", True)
-            self.update_layer(adapter_name, r, lora_alpha, lora_dropout, init_lora_weights)
-            self.active_adapter = adapter_name
-
-        def forward(self, x: torch.Tensor) -> torch.Tensor:
-            # note: logic differs from default Linear because merging is not supported
-            result = super().forward(x)
-
-            if (
-                self.disable_adapters
-                or (self.active_adapter not in self.lora_A.keys())
-                or (self.r[self.active_adapter] == 0)
-            ):
-                return result
-
-            lora_A = self.lora_A[self.active_adapter]
-            lora_B = self.lora_B[self.active_adapter]
-            dropout = self.lora_dropout[self.active_adapter]
-            scaling = self.scaling[self.active_adapter]
-
-            requires_conversion = not torch.is_autocast_enabled()
-            if requires_conversion:
-                expected_dtype = result.dtype
-                if x.dtype != torch.float32:
-                    x = x.float()
-
-            output = lora_B(lora_A(dropout(x)))
-            if requires_conversion:
-                output = output.to(expected_dtype)
-            output = output * scaling
-            result += output
-            return result
-
-
-if is_bnb_4bit_available():
-
-    class Linear4bit(bnb.nn.Linear4bit, LoraLayer):
-        # Lora implemented in a dense layer
-        def __init__(
-            self,
-            adapter_name,
-            in_features,
-            out_features,
-            r: int = 0,
-            lora_alpha: int = 1,
-            lora_dropout: float = 0.0,
-            **kwargs,
-        ) -> None:
-            bnb.nn.Linear4bit.__init__(
-                self,
-                in_features,
-                out_features,
-                bias=kwargs.get("bias", True),
-                compute_dtype=kwargs.get("compute_dtype", torch.float32),
-                compress_statistics=kwargs.get("compress_statistics", True),
-                quant_type=kwargs.get("quant_type", "nf4"),
-            )
-            LoraLayer.__init__(self, in_features=in_features, out_features=out_features)
-
-            # Freezing the pre-trained weight matrix
-            self.weight.requires_grad = False
-
-            init_lora_weights = kwargs.pop("init_lora_weights", True)
-            self.update_layer(adapter_name, r, lora_alpha, lora_dropout, init_lora_weights)
-            self.active_adapter = adapter_name
-
-        def forward(self, x: torch.Tensor) -> torch.Tensor:
-            # note: logic differs from default Linear because merging is not supported
-            result = super().forward(x)
-
-            if (
-                self.disable_adapters
-                or (self.active_adapter not in self.lora_A.keys())
-                or (self.r[self.active_adapter] == 0)
-            ):
-                return result
-
-            # As per Tim Dettmers, for 4bit, we need to defensively clone here.
-            # The reason is that in some cases, an error can occur that backprop
-            # does not work on a manipulated view. This issue may be solved with
-            # newer PyTorch versions but this would need extensive testing to be
-            # sure.
-            result = result.clone()
-
-            lora_A = self.lora_A[self.active_adapter]
-            lora_B = self.lora_B[self.active_adapter]
-            dropout = self.lora_dropout[self.active_adapter]
-            scaling = self.scaling[self.active_adapter]
-
-            requires_conversion = not torch.is_autocast_enabled()
-            if requires_conversion:
-                expected_dtype = result.dtype
-                x = x.to(lora_A.weight.dtype)
-
-            output = lora_B(lora_A(dropout(x)))
-            if requires_conversion:
-                output = output.to(expected_dtype)
-            output = output * scaling
-            result += output
-            return result
-
-
-class QuantLinear(torch.nn.Module, LoraLayer):
-    def __init__(
-        self,
-        adapter_name,
-        quant_linear_module,
-        r: int = 0,
-        lora_alpha: int = 1,
-        lora_dropout: float = 0.0,
-        **kwargs,
-    ):
-        torch.nn.Module.__init__(self)
-        LoraLayer.__init__(
-            self, in_features=quant_linear_module.infeatures, out_features=quant_linear_module.outfeatures
-        )
-        self.quant_linear_module = quant_linear_module
-        self.weight = quant_linear_module.qweight
-        init_lora_weights = kwargs.pop("init_lora_weights", True)
-        self.update_layer(adapter_name, r, lora_alpha, lora_dropout, init_lora_weights)
-        self.active_adapter = adapter_name
-
-    def forward(self, x: torch.Tensor):
-        # note: logic differs from default Linear because merging is not supported
-        result = self.quant_linear_module(x)
-
-        if (
-            self.disable_adapters
-            or (self.active_adapter not in self.lora_A.keys())
-            or (self.r[self.active_adapter] == 0)
-        ):
-            return result
-
-        lora_A = self.lora_A[self.active_adapter]
-        lora_B = self.lora_B[self.active_adapter]
-        dropout = self.lora_dropout[self.active_adapter]
-        scaling = self.scaling[self.active_adapter]
-
-        requires_conversion = not torch.is_autocast_enabled()
-        if requires_conversion:
-            expected_dtype = result.dtype
-            x = x.to(lora_A.weight.dtype)
-
-        output = lora_B(lora_A(dropout(x)))
-        if requires_conversion:
-            output = output.to(expected_dtype)
-        output = output * scaling
-        result += output
-        return result
-
-    # TODO: Check if it is better as suggested by users https://github.com/PanQiWei/AutoGPTQ/pull/102
-    # def reset_lora_parameters(self, adapter_name):
-    #     if adapter_name in self.lora_A.keys():
-    #         torch.nn.init.xavier_uniform_(self.lora_A[adapter_name].weight)
-    #         torch.nn.init.zeros_(self.lora_B[adapter_name].weight)
