+++ src/peft/tuners/prefix_tuning/__init__.py
@@ -13,23 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from peft.import_utils import is_bnb_4bit_available, is_bnb_available
-
-from .config import LoraConfig
-from .gptq import QuantLinear
-from .layer import Conv2d, Embedding, Linear, LoraLayer
-from .model import LoraModel
+from .config import PrefixTuningConfig
+from .model import PrefixEncoder
 
 
-__all__ = ["LoraConfig", "Conv2d", "Embedding", "LoraLayer", "Linear", "LoraModel", "QuantLinear"]
-
-
-if is_bnb_available():
-    from .bnb import Linear8bitLt
-
-    __all__ += ["Linear8bitLt"]
-
-if is_bnb_4bit_available():
-    from .bnb import Linear4bit
-
-    __all__ += ["Linear4bit"]
+__all__ = ["PrefixTuningConfig", "PrefixEncoder"]
