+++ src/peft/tuners/prompt_tuning/model.py
@@ -13,54 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import enum
 import math
-from dataclasses import dataclass, field
-from typing import Optional, Union
 
 import torch
 
-from ..config import PromptLearningConfig
-from ..utils import PeftType
-
-
-class PromptTuningInit(str, enum.Enum):
-    TEXT = "TEXT"
-    RANDOM = "RANDOM"
-
-
-@dataclass
-class PromptTuningConfig(PromptLearningConfig):
-    """
-    This is the configuration class to store the configuration of a [`PromptEmbedding`].
-
-    Args:
-        prompt_tuning_init (Union[[`PromptTuningInit`], `str`]): The initialization of the prompt embedding.
-        prompt_tuning_init_text (`str`, *optional*):
-            The text to initialize the prompt embedding. Only used if `prompt_tuning_init` is `TEXT`.
-        tokenizer_name_or_path (`str`, *optional*):
-            The name or path of the tokenizer. Only used if `prompt_tuning_init` is `TEXT`.
-    """
-
-    prompt_tuning_init: Union[PromptTuningInit, str] = field(
-        default=PromptTuningInit.RANDOM,
-        metadata={"help": "How to initialize the prompt tuning parameters"},
-    )
-    prompt_tuning_init_text: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "The text to use for prompt tuning initialization. Only used if prompt_tuning_init is `TEXT`"
-        },
-    )
-    tokenizer_name_or_path: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "The tokenizer to use for prompt tuning initialization. Only used if prompt_tuning_init is `TEXT`"
-        },
-    )
-
-    def __post_init__(self):
-        self.peft_type = PeftType.PROMPT_TUNING
+from .config import PromptTuningInit
 
 
 class PromptEmbedding(torch.nn.Module):
